feat: improve overall spacing and padding for post body in condensed/expanded views
